Fix duplicated description
